pyproject.toml
[tool.poetry]
name = "wf-camera-utils"
version = "1.0.28"
description = "utilities for wildflower cameras and raspberry pi"
authors = ["Paul J DeCoursey <paul@decoursey.net>"]
license = "MIT"
readme = "README.md"
packages = [{include = "capture"}, {include = "capture_v2"}, {include = "cuwb_stream"}]


[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"
black = "^22.10.0"


[tool.poetry.group.capture.dependencies]
picamera = {version = "1.13", platform = "linux"}
ffmpeg-python = "0.2.0"
minio = ">=7.1.12"
pyyaml = ">=6.0"
python-json-logger = ">=2.0.4"
influx_line_protocol = ">=0.1.5"


[tool.poetry.group.capture_v2.dependencies]
picamera2 = {version = "^0.3.9", platform = "linux"}
pandas = "^2.0.1"
pydantic = "^1.10.7"
python-prctl = {version = "^1.8.1", platform = "linux"}


[tool.poetry.group.cuwb_stream.dependencies]
cdp-py = "1.5.0"
click = ">=8.1.3"
requests = ">=2.28.1"
tenacity = ">=8.1.0"
apscheduler = ">=3.10.1"


[tool.poetry.scripts]
cuwb_stream = "cuwb_stream.tools.__main__:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
